requirements.txt
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
pandas==1.3.3
plotly==5.3.1
nltk==3.6.2
openai==0.27.0
python-dotenv==0.19.1
praw==7.4.0
requests==2.26.0
spacy==3.1.3
tqdm==4.62.3